backend/requirements.txt
loguru==0.7.2
websockets==13.1
streamcontroller-plugin-tools>=2.0.1
PyJWT==2.9.0
cryptography>=41.0.0
